+++ jrasp-daemon/bin/service.sh
@@ -2,12 +2,13 @@
 
 while true
 do
-  process=`ps aux | grep jrasp-daemon | grep -v grep`;
-  if [ "$process" == "" ]; then
-      echo "jrasp-daemon is not running";
-      ./jrasp-daemon;
+  ## issue 41
+  process=$(ps aux | grep -E "jrasp-daemon$" | grep -v grep)
+  if [ -z "$process" ]; then
+    echo "jrasp-daemon is not running"
+    ./jrasp-daemon
   else
-      sleep 10;
-      echo "jrasp-daemon is running";
+    sleep 10
+    echo "jrasp-daemon is running"
   fi
 done