+++ jrasp-maven/src/main/java/org/eclipse/jetty/server/Response.java
@@ -1,4 +1,7 @@
 package org.eclipse.jetty.server;
 
 public class Response {
+    public void addHeader(String name, String value) {
+
+    }
 }