+++ jrasp-maven/src/main/java/org/sparkproject/jetty/server/Response.java
@@ -1,5 +1,7 @@
 package org.sparkproject.jetty.server;
 
 public class Response {
+    public void addHeader(String name, String value) {
 
+    }
 }